Enable sending transactions without txindex=1
This fixes a bug in the calls to wallet RPCs while fetching raw
transaction hex of UTXOs that are selected as part of an outgoing
transaction. The hex was always empty since the DecodeRawTransaction
method of btcd never populates the original hex in the TxRawResult
struct. A fix may be upstreamed to btcd in future.

Closes #23: Impossible to build UTXO map without txindex=1
  The system does not rely on a complete UTXO set, and txindex=1 is no
  longer a requirement. If txindex is not available, the Vin component
  of incoming transactions would have partial information (ex, missing
  address and value fields), but it won't be required by libcore to
  synchronize the accounts properly, or prepare outgoing transactions.

Closes #15: Drop usage of getrawtransaction RPC
  getrawtransaction RPC is now optional, and only used when txindex is
  enabled. If txindex is not available, we fall back to wallet RPCs,
  which is marginally slower than a call to getrawtransaction.