Implement stateless status monitoring
Remove the Status field from the Bus struct, and compute it on-demand
through the /blockchain/v3/explorer/status endpoint. The improved
monitoring system eliminates the possibility to have inconsistent
status.

/!\ The following statuses have been removed:
- Initializing
- PendingScan

Added a new status NodeDisconnected to indicate that bitcoind is
unreachable.

The new lifecyle of happy-path statuses is as follows:
- Syncing
- Scanning
- Ready